+++ reflect/src/main/java/com/github/chevyself/reflect/debug/LogTask.java
@@ -4,11 +4,20 @@ import com.github.chevyself.reflect.debug.tasks.DebugTask;
 import lombok.Getter;
 import lombok.NonNull;
 
+/**
+ * Represents a task that has messages to be logged.
+ */
 public class LogTask implements DebugTask {
 
   @Getter protected final int id;
   @Getter protected final Debugger debugger;
 
+  /**
+   * Creates a new log task.
+   *
+   * @param id the id of the task
+   * @param debugger the debugger that created the task
+   */
   public LogTask(int id, Debugger debugger) {
     this.id = id;
     this.debugger = debugger;