+++ reflect/src/main/java/com/github/chevyself/reflect/modifiers/Modifier.java
@@ -3,6 +3,7 @@ package com.github.chevyself.reflect.modifiers;
 import com.github.chevyself.reflect.wrappers.WrappedField;
 import java.lang.reflect.InvocationTargetException;
 import lombok.NonNull;
+import org.jetbrains.annotations.Nullable;
 
 /**
  * Implement this class to change the value of a field. using {@link WrappedField#set(Object,
@@ -11,7 +12,7 @@ import lombok.NonNull;
 public interface Modifier {
 
   /**
-   * Modify a field.
+   * Modifies a field.
    *
    * @param field the field to be modified
    * @param reference the reference in which the field will be changed
@@ -20,6 +21,6 @@ public interface Modifier {
    * @throws IllegalAccessException if this Field object is enforcing Java language access control
    *     and the underlying field is either inaccessible or final.
    */
-  boolean modify(@NonNull WrappedField<?> field, Object reference)
+  boolean modify(@NonNull WrappedField<?> field, @Nullable Object reference)
       throws IllegalAccessException, InvocationTargetException;
 }