+++ CHANGELOG.md
@@ -18,4 +18,5 @@ Changelog
 - [X] Simplify boss-bars, instead of having two methods as `Channel#getBossBar` and `Channel#giveBossBar` now there is only one method `Channel#getBossBar` which returns one that can be displayed.
   - [X] Boss bars must be displayed using `AdaptedBossBar#display`
 - [X] Add support to `Packet`s which changed in Minecraft 1.17
+- [X] Better `Reflect` debugging
 - Moved `Versions#getObjectiveAdapter` to `Players#getObjectiveAdapter`